+++ src/main/java/net/geforcemods/securitycraft/screen/BlockPocketManagerScreen.java
@@ -8,16 +8,16 @@ import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.blockentities.BlockPocketManagerBlockEntity;
 import net.geforcemods.securitycraft.inventory.BlockPocketManagerMenu;
+import net.geforcemods.securitycraft.network.server.AssembleBlockPocket;
 import net.geforcemods.securitycraft.network.server.SyncBlockPocketManager;
+import net.geforcemods.securitycraft.network.server.ToggleBlockPocketManager;
 import net.geforcemods.securitycraft.screen.components.CallbackSlider;
 import net.geforcemods.securitycraft.screen.components.ColorChooser;
 import net.geforcemods.securitycraft.screen.components.ColorChooserButton;
 import net.geforcemods.securitycraft.screen.components.StackHoverChecker;
 import net.geforcemods.securitycraft.screen.components.ToggleComponentButton;
 import net.geforcemods.securitycraft.util.IHasExtraAreas;
-import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.geforcemods.securitycraft.util.Utils;
-import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Button;
@@ -25,7 +25,6 @@ import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.Rect2i;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.ClickType;
@@ -114,7 +113,7 @@ public class BlockPocketManagerScreen extends AbstractContainerScreen<BlockPocke
 			sizeButton.active = toggleButton.active = assembleButton.active = outlineButton.active = offsetSlider.active = colorChooserButton.active = false;
 		else {
 			updateMaterialInformation(true);
-			sizeButton.active = offsetSlider.active = !be.isEnabled();
+			sizeButton.active = assembleButton.active = offsetSlider.active = !be.isEnabled();
 		}
 
 		if (!storage) {
@@ -281,18 +280,9 @@ public class BlockPocketManagerScreen extends AbstractContainerScreen<BlockPocke
 	}
 
 	public void toggleButtonClicked(Button button) {
-		if (be.isEnabled())
-			be.disableMultiblock();
-		else {
-			MutableComponent feedback;
-
-			be.setSize(size);
-			feedback = be.enableMultiblock();
-
-			if (feedback != null)
-				PlayerUtils.sendMessageToPlayer(Minecraft.getInstance().player, Utils.localize(SCContent.BLOCK_POCKET_MANAGER.get().getDescriptionId()), feedback, ChatFormatting.DARK_AQUA, true);
-		}
-
+		be.setSize(size);
+		be.setEnabled(!be.isEnabled());
+		SecurityCraft.CHANNEL.send(PacketDistributor.SERVER.noArg(), new ToggleBlockPocketManager(be, be.isEnabled()));
 		Minecraft.getInstance().player.closeContainer();
 	}
 
@@ -325,14 +315,8 @@ public class BlockPocketManagerScreen extends AbstractContainerScreen<BlockPocke
 	}
 
 	public void assembleButtonClicked(Button button) {
-		MutableComponent feedback;
-
 		be.setSize(size);
-		feedback = be.autoAssembleMultiblock();
-
-		if (feedback != null)
-			PlayerUtils.sendMessageToPlayer(Minecraft.getInstance().player, Utils.localize(SCContent.BLOCK_POCKET_MANAGER.get().getDescriptionId()), feedback, ChatFormatting.DARK_AQUA, true);
-
+		SecurityCraft.CHANNEL.send(PacketDistributor.SERVER.noArg(), new AssembleBlockPocket(be));
 		Minecraft.getInstance().player.closeContainer();
 	}